Updated the stream configuration to also publish to a topic
Now the stream in MyKafkaStreamsConfiguration also publishes to the topic. This allows to delete the dead code that was considered to maybe accomplish the same goal. Besides, log were removed.

The solution is still a bit brittle now - the dependency in WordCountResource is manually set at runtime when the stream is created. This means that each call made to the API while the stream is not ready will fail. This is not desirable and should be improved.